@@ -1,8 +1,6 @@
 // Copyright (c) 2018, Juan Mellado. All rights reserved. Use of this source
 // is governed by a MIT-style license that can be found in the LICENSE file.
 
-// @dart=2.9
-
 import 'package:test/test.dart';
 
 // ignore: directives_ordering
@@ -11,8 +9,8 @@ import 'package:dartis/dartis.dart';
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  PubSubCommands<String, String> commands;
+  late Client client;
+  late PubSubCommands<String?, String> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
@@ -60,8 +58,8 @@ void main() {
     group('support', () {
       group('PubsubResult', () {
         test('toString', () {
-          const value = PubsubResult<String>(null, null);
-          expect(value.toString(), startsWith('PubsubResult<String>:'));
+          const value = PubsubResult<String?>(null, null);
+          expect(value.toString(), startsWith('PubsubResult<String?>:'));
         });
       });
     });
