@@ -1,8 +1,6 @@
 // Copyright (c) 2018, Juan Mellado. All rights reserved. Use of this source
 // is governed by a MIT-style license that can be found in the LICENSE file.
 
-// @dart=2.9
-
 import 'package:test/test.dart';
 
 // ignore: directives_ordering
@@ -11,8 +9,8 @@ import 'package:dartis/dartis.dart';
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  SortedSetCommands<String, String> commands;
+  late Client client;
+  late SortedSetCommands<String?, String?> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
@@ -33,20 +31,20 @@ void main() {
 
       // Pop from one sorted set.
       var result = await commands.bzpopmax(key: key1);
-      expect(result.key, equals(key1));
-      expect(result.member.key, equals('b'));
-      expect(result.member.value, equals(2.0));
+      expect(result!.key, equals(key1));
+      expect(result.member!.key, equals('b'));
+      expect(result.member!.value, equals(2.0));
 
       result = await commands.bzpopmax(key: key1);
-      expect(result.key, equals(key1));
-      expect(result.member.key, equals('a'));
-      expect(result.member.value, equals(1.0));
+      expect(result!.key, equals(key1));
+      expect(result.member!.key, equals('a'));
+      expect(result.member!.value, equals(1.0));
 
       // Pop from some sorted sets.
       result = await commands.bzpopmax(keys: [key1, key2]);
-      expect(result.key, equals(key2));
-      expect(result.member.key, equals('c'));
-      expect(result.member.value, equals(3.0));
+      expect(result!.key, equals(key2));
+      expect(result.member!.key, equals('c'));
+      expect(result.member!.value, equals(3.0));
 
       // Pop blocking.
       expect(await commands.bzpopmax(key: key1, timeout: 1), isNull);
@@ -66,22 +64,22 @@ void main() {
       // Pop from one sorted set.
       var result = await commands.bzpopmin(key: key1);
       expect(result, isNotNull);
-      expect(result.key, equals(key1));
-      expect(result.member.key, equals('a'));
-      expect(result.member.value, equals(1.0));
+      expect(result!.key, equals(key1));
+      expect(result.member!.key, equals('a'));
+      expect(result.member!.value, equals(1.0));
 
       result = await commands.bzpopmin(key: key1);
       expect(result, isNotNull);
-      expect(result.key, equals(key1));
-      expect(result.member.key, equals('b'));
-      expect(result.member.value, equals(2.0));
+      expect(result!.key, equals(key1));
+      expect(result.member!.key, equals('b'));
+      expect(result.member!.value, equals(2.0));
 
       // Pop from one from some sorted sets.
       result = await commands.bzpopmin(keys: [key1, key2]);
       expect(result, isNotNull);
-      expect(result.key, equals(key2));
-      expect(result.member.key, equals('c'));
-      expect(result.member.value, equals(3.0));
+      expect(result!.key, equals(key2));
+      expect(result.member!.key, equals('c'));
+      expect(result.member!.value, equals(3.0));
 
       // Pop blocking.
       expect(await commands.bzpopmin(key: key1, timeout: 1), isNull);
@@ -597,9 +595,9 @@ void main() {
 
       group('SortedSetPopResult', () {
         test('toString', () {
-          const value = SortedSetPopResult<String, String>(null, null);
+          const value = SortedSetPopResult<String?, String>(null, null);
           expect(value.toString(),
-              startsWith('SortedSetPopResult<String, String>:'));
+              startsWith('SortedSetPopResult<String?, String>:'));
         });
       });
 
