@@ -8,7 +8,7 @@ import 'command.dart';
 import 'module.dart';
 
 /// Redis commands.
-class Commands<K, V> extends ModuleBase
+class Commands<K extends Object, V extends Object> extends ModuleBase
     implements
         ClusterCommands<K>,
         ConnectionCommands,
@@ -32,89 +32,87 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<void> clusterAddslots({int? slot, Iterable<int> slots = const []}) =>
-      run<void>(<Object?>[r'CLUSTER', r'ADDSLOTS', slot, ...slots]);
+      run<Never>(<Object?>[r'CLUSTER', r'ADDSLOTS', slot, ...slots]);
 
   @override
   Future<int?> clusterCountFailureReports(String nodeId) =>
-      run<int?>(<Object>[r'CLUSTER', r'COUNT-FAILURE-REPORTS', nodeId]);
+      run<int>(<Object>[r'CLUSTER', r'COUNT-FAILURE-REPORTS', nodeId]);
 
   @override
   Future<int?> clusterCountkeysinslot(int slot) =>
-      run<int?>(<Object>[r'CLUSTER', r'COUNTKEYSINSLOT', slot]);
+      run<int>(<Object>[r'CLUSTER', r'COUNTKEYSINSLOT', slot]);
 
   @override
   Future<void> clusterDelslots({int? slot, Iterable<int> slots = const []}) =>
-      run<void>(<Object?>[r'CLUSTER', r'DELSLOTS', slot, ...slots]);
+      run<Never>(<Object?>[r'CLUSTER', r'DELSLOTS', slot, ...slots]);
 
   @override
   Future<void> clusterFailover([ClusterFailoverMode? mode]) =>
-      run<void>(<Object?>[r'CLUSTER', r'FAILOVER', mode?.name]);
+      run<Never>(<Object?>[r'CLUSTER', r'FAILOVER', mode?.name]);
 
   @override
   Future<void> clusterForget(String nodeId) =>
-      run<void>(<Object>[r'CLUSTER', r'FORGET', nodeId]);
+      run<Never>(<Object>[r'CLUSTER', r'FORGET', nodeId]);
 
   @override
   Future<List<K>?> clusterGetkeysinslot(int slot, int count) =>
-      run<List<K>?>(<Object>[r'CLUSTER', r'GETKEYSINSLOT', slot, count]);
+      run<List<K>>(<Object>[r'CLUSTER', r'GETKEYSINSLOT', slot, count]);
 
   @override
   Future<Map<String, String>?> clusterInfo() =>
-      run<Map<String, String>?>(<Object>[r'CLUSTER', r'INFO'],
+      run<Map<String, String>>(<Object>[r'CLUSTER', r'INFO'],
           mapper: clusterInfoMapper);
 
   @override
   Future<int?> clusterKeyslot(K key) =>
-      run<int?>(<Object?>[r'CLUSTER', r'KEYSLOT', key]);
+      run<int>(<Object?>[r'CLUSTER', r'KEYSLOT', key]);
 
   @override
   Future<void> clusterMeet(String ip, int port) =>
-      run<void>(<Object>[r'CLUSTER', r'MEET', ip, port]);
+      run<Never>(<Object>[r'CLUSTER', r'MEET', ip, port]);
 
   @override
-  Future<String?> clusterNodes() =>
-      run<String?>(<Object>[r'CLUSTER', r'NODES']);
+  Future<String?> clusterNodes() => run<String>(<Object>[r'CLUSTER', r'NODES']);
 
   @override
   Future<void> clusterReplicate(String nodeId) =>
-      run<void>(<Object>[r'CLUSTER', r'REPLICATE', nodeId]);
+      run<Never>(<Object>[r'CLUSTER', r'REPLICATE', nodeId]);
 
   @override
   Future<void> clusterReset([ClusterResetMode? mode]) =>
-      run<void>(<Object?>[r'CLUSTER', r'RESET', mode?.name]);
+      run<Never>(<Object?>[r'CLUSTER', r'RESET', mode?.name]);
 
   @override
-  Future<void> clusterSaveconfig() =>
-      run<void>(<Object>[r'CLUSTER', r'SAVECONFIG']);
+  Future<void> clusterSaveconfig() => run<Never>(<Object>[r'CLUSTER', r'SAVECONFIG']);
 
   @override
   Future<void> clusterSetConfigEpoch(int configEpoch) =>
-      run<void>(<Object>[r'CLUSTER', r'SET-CONFIG-EPOCH', configEpoch]);
+      run<Never>(<Object>[r'CLUSTER', r'SET-CONFIG-EPOCH', configEpoch]);
 
   @override
   Future<void> clusterSetslot(int slot, ClusterSetslotCommand command,
           {String? nodeId}) =>
-      run<void>(<Object?>[r'CLUSTER', r'SETSLOT', slot, command.name, nodeId]);
+      run<Never>(<Object?>[r'CLUSTER', r'SETSLOT', slot, command.name, nodeId]);
 
   @override
   Future<String?> clusterSlaves(String nodeId) =>
-      run<String?>(<Object>[r'CLUSTER', r'SLAVES', nodeId]);
+      run<String>(<Object>[r'CLUSTER', r'SLAVES', nodeId]);
 
   @override
   Future<List<ClusterSlotRange>?> clusterSlots() =>
-      run<List<ClusterSlotRange>?>(<Object>[r'CLUSTER', r'SLOTS'],
+      run<List<ClusterSlotRange>>(<Object>[r'CLUSTER', r'SLOTS'],
           mapper: clusterSlotRangeMapper);
 
   @override
-  Future<void> readonly() => run<void>(<Object>[r'READONLY']);
+  Future<void> readonly() => run<Never>(<Object>[r'READONLY']);
 
   @override
-  Future<void> readwrite() => run<void>(<Object>[r'READWRITE']);
+  Future<void> readwrite() => run<Never>(<Object>[r'READWRITE']);
 
   // Connection.
 
   @override
-  Future<void> auth(String password) => run<void>(<Object>[r'AUTH', password]);
+  Future<void> auth(String password) => run<Never>(<Object>[r'AUTH', password]);
 
   @override
   Future<String?> echo(String? message) =>
@@ -125,21 +123,21 @@ class Commands<K, V> extends ModuleBase
       run<String>(<Object?>[r'PING', message]);
 
   @override
-  Future<void> quit() => run<void>(<Object>[r'QUIT']);
+  Future<void> quit() => run<Never>(<Object>[r'QUIT']);
 
   @override
-  Future<void> select(int index) => run<void>(<Object>[r'SELECT', index]);
+  Future<void> select(int index) => run<Never>(<Object>[r'SELECT', index]);
 
   @override
   Future<void> swapdb(int index1, int index2) =>
-      run<void>(<Object>[r'SWAPDB', index1, index2]);
+      run<Never>(<Object>[r'SWAPDB', index1, index2]);
 
   // Geo.
 
   @override
   Future<int?> geoadd(K key,
           {GeoItem<V>? item, Iterable<GeoItem<V>> items = const []}) =>
-      run<int?>(<Object?>[
+      run<int>(<Object?>[
         r'GEOADD',
         key,
         ..._expandGeoItem(item),
@@ -152,17 +150,17 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<double?> geodist(K key, V member1, V member2, {GeoUnit? unit}) =>
-      run<double?>(<Object?>[r'GEODIST', key, member1, member2, unit?.name]);
+      run<double>(<Object?>[r'GEODIST', key, member1, member2, unit?.name]);
 
   @override
   Future<List<String>?> geohash(K key,
           {V? member, Iterable<V> members = const []}) =>
-      run<List<String>?>(<Object?>[r'GEOHASH', key, member, ...members]);
+      run<List<String>>(<Object?>[r'GEOHASH', key, member, ...members]);
 
   @override
   Future<List<GeoPosition?>?> geopos(K key,
           {V? member, Iterable<V> members = const []}) =>
-      run<List<GeoPosition?>?>(<Object?>[r'GEOPOS', key, member, ...members],
+      run<List<GeoPosition?>>(<Object?>[r'GEOPOS', key, member, ...members],
           mapper: geoPositionMapper);
 
   @override
@@ -173,7 +171,7 @@ class Commands<K, V> extends ModuleBase
           bool withHash = false,
           int? count,
           GeoOrder? order}) =>
-      run<List<GeoradiusResult<V>>?>(<Object?>[
+      run<List<GeoradiusResult<V>>>(<Object?>[
         r'GEORADIUS_RO',
         key,
         longitude,
@@ -194,7 +192,7 @@ class Commands<K, V> extends ModuleBase
   Future<int?> georadiusStore(
           K key, double longitude, double latitude, double radius, GeoUnit unit,
           {int? count, GeoOrder? order, K? storeKey, K? storeDistKey}) =>
-      run<int?>(<Object?>[
+      run<int>(<Object?>[
         r'GEORADIUS',
         key,
         longitude,
@@ -218,7 +216,7 @@ class Commands<K, V> extends ModuleBase
           bool withHash = false,
           int? count,
           GeoOrder? order}) =>
-      run<List<GeoradiusResult<V>>?>(<Object?>[
+      run<List<GeoradiusResult<V>>>(<Object?>[
         r'GEORADIUSBYMEMBER_RO',
         key,
         member,
@@ -238,7 +236,7 @@ class Commands<K, V> extends ModuleBase
   Future<int?> georadiusbymemberStore(
           K key, V member, double radius, GeoUnit unit,
           {int? count, GeoOrder? order, K? storeKey, K? storeDistKey}) =>
-      run<int?>(<Object?>[
+      run<int>(<Object?>[
         r'GEORADIUSBYMEMBER',
         key,
         member,
@@ -257,40 +255,40 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int?> hdel(K key, {K? field, Iterable<K> fields = const []}) =>
-      run<int?>(<Object?>[r'HDEL', key, field, ...fields]);
+      run<int>(<Object?>[r'HDEL', key, field, ...fields]);
 
   @override
   Future<int?> hexists(K key, K field) =>
-      run<int?>(<Object?>[r'HEXISTS', key, field]);
+      run<int>(<Object?>[r'HEXISTS', key, field]);
 
   @override
-  Future<V?> hget(K key, K field) => run<V?>(<Object?>[r'HGET', key, field]);
+  Future<V?> hget(K key, K field) => run<V>(<Object?>[r'HGET', key, field]);
 
   @override
   Future<Map<K, V?>?> hgetall(K key) =>
-      run<Map<K, V?>?>(<Object?>[r'HGETALL', key], mapper: HashMapper<K, V?>());
+      run<Map<K, V?>>(<Object?>[r'HGETALL', key], mapper: HashMapper<K, V>());
 
   @override
   Future<int?> hincrby(K key, K field, int increment) =>
-      run<int?>(<Object?>[r'HINCRBY', key, field, increment]);
+      run<int>(<Object?>[r'HINCRBY', key, field, increment]);
 
   @override
   Future<double?> hincrbyfloat(K key, K field, double increment) =>
-      run<double?>(<Object?>[r'HINCRBYFLOAT', key, field, increment]);
+      run<double>(<Object?>[r'HINCRBYFLOAT', key, field, increment]);
 
   @override
-  Future<List<K>?> hkeys(K key) => run<List<K>?>(<Object?>[r'HKEYS', key]);
+  Future<List<K>?> hkeys(K key) => run<List<K>>(<Object?>[r'HKEYS', key]);
 
   @override
-  Future<int?> hlen(K key) => run<int?>(<Object?>[r'HLEN', key]);
+  Future<int?> hlen(K key) => run<int>(<Object?>[r'HLEN', key]);
 
   @override
-  Future<List<V>?> hmget(K key, {K? field, Iterable<K> fields = const []}) =>
-      run<List<V>?>(<Object?>[r'HMGET', key, field, ...fields]);
+  Future<List<V?>> hmget(K key, {K? field, Iterable<K> fields = const []}) =>
+      runNonNull<List<V?>>(<Object?>[r'HMGET', key, field, ...fields]);
 
   @override
   Future<void> hmset(K key, {K? field, V? value, Map<K, V?> hash = const {}}) =>
-      run<void>(<Object?>[
+      run<Never>(<Object?>[
         r'HMSET',
         key,
         field,
@@ -301,7 +299,7 @@ class Commands<K, V> extends ModuleBase
   @override
   Future<HashScanResult<K, V?>?> hscan(K key, int cursor,
           {K? pattern, int? count}) =>
-      run<HashScanResult<K, V?>?>(<void>[
+      run<HashScanResult<K, V?>>(<void>[
         r'HSCAN',
         key,
         cursor,
@@ -309,61 +307,63 @@ class Commands<K, V> extends ModuleBase
         pattern,
         count == null ? null : r'COUNT',
         count
-      ], mapper: HashScanMapper<K, V?>());
+      ], mapper: HashScanMapper<K, V>());
 
   @override
   Future<int?> hset(K key, K field, V? value) =>
-      run<int?>(<Object?>[r'HSET', key, field, value]);
+      run<int>(<Object?>[r'HSET', key, field, value]);
 
   @override
   Future<int?> hsetnx(K key, K field, V? value) =>
-      run<int?>(<Object?>[r'HSETNX', key, field, value]);
+      run<int>(<Object?>[r'HSETNX', key, field, value]);
 
   @override
   Future<int?> hstrlen(K key, K field) =>
-      run<int?>(<Object?>[r'HSTRLEN', key, field]);
+      run<int>(<Object?>[r'HSTRLEN', key, field]);
 
   @override
-  Future<List<V?>?> hvals(K? key) => run<List<V?>?>(<Object?>[r'HVALS', key]);
+  Future<List<V?>?> hvals(K? key) => run<List<V?>>(<Object?>[r'HVALS', key]);
 
   // HyperLogLogs.
 
   @override
   Future<int?> pfadd(K key, {V? element, Iterable<V> elements = const []}) =>
-      run<int?>(<Object?>[r'PFADD', key, element, ...elements]);
+      run<int>(<Object?>[r'PFADD', key, element, ...elements]);
 
   @override
   Future<int?> pfcount({K? key, Iterable<K> keys = const []}) =>
-      run<int?>(<Object?>[r'PFCOUNT', key, ...keys]);
+      run<int>(<Object?>[r'PFCOUNT', key, ...keys]);
 
   @override
   Future<void> pfmerge(K destkey,
           {K? sourcekey, Iterable<K> sourcekeys = const []}) =>
-      run<void>(<Object?>[r'PFMERGE', destkey, sourcekey, ...sourcekeys]);
+      run<Never>(<Object?>[r'PFMERGE', destkey, sourcekey, ...sourcekeys]);
 
   // Keys.
 
   @override
   Future<int> del({K? key, Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'DEL', key, ...keys]);
+      runNonNull<int>(<Object>[r'DEL', if (key != null) key, ...keys]);
 
   @override
-  Future<List<int>> dump(K key) => run<List<int>>(<Object?>[r'DUMP', key]);
+  Future<List<int>?> dump(K key) =>
+      run<List<int>>(<Object>[r'DUMP', key]);
 
   @override
   Future<int> exists({K? key, Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'EXISTS', key, ...keys]);
+      runNonNull<int>(<Object>[r'EXISTS', if (key != null) key, ...keys]);
 
   @override
   Future<int> expire(K key, int seconds) =>
-      run<int>(<Object?>[r'EXPIRE', key, seconds]);
+      runNonNull<int>(<Object>[r'EXPIRE', key, seconds]);
 
   @override
   Future<int> expireat(K key, int timestamp) =>
-      run<int>(<Object?>[r'EXPIREAT', key, timestamp]);
+      runNonNull<int>(<Object>[r'EXPIREAT', key, timestamp]);
 
   @override
-  Future<List<K>> keys(K pattern) => run<List<K>>(<Object?>[r'KEYS', pattern]);
+  Future<List<K>> keys(K pattern) =>
+      runNonNull<List<K>>(<Object>[r'KEYS', pattern]);
 
   @override
   Future<String> migrate(String host, int port, int destinationDb, int timeout,
@@ -371,59 +371,60 @@ class Commands<K, V> extends ModuleBase
           bool replace = false,
           K? key,
           Iterable<K> keys = const []}) =>
-      run<String>(<Object?>[
+      runNonNull<String>(<Object>[
         r'MIGRATE',
         host,
         port,
         key ?? r'',
         destinationDb,
         timeout,
-        copy ? r'COPY' : null,
-        replace ? r'REPLACE' : null,
-        keys.isNotEmpty ? r'KEYS' : null,
+        if (copy) r'COPY',
+        if (replace) r'REPLACE',
+        if (keys.isNotEmpty) r'KEYS',
         ...keys
       ]);
 
   @override
-  Future<int> move(K key, int db) => run<int>(<Object?>[r'MOVE', key, db]);
+  Future<int> move(K key, int db) =>
+      runNonNull<int>(<Object>[r'MOVE', key, db]);
 
   @override
-  Future<String> object(ObjectSubcommand subcommand, K key) =>
-      run<String>(<Object?>[r'OBJECT', subcommand.name, key]);
+  Future<String?> object(ObjectSubcommand subcommand, K key) =>
+      run<String>(<Object>[r'OBJECT', subcommand.name, key]);
 
   @override
   Future<List<String>> objectHelp() =>
-      run<List<String>>(<Object>[r'OBJECT', r'HELP']);
+      runNonNull<List<String>>(<Object>[r'OBJECT', r'HELP']);
 
   @override
-  Future<int> persist(K key) => run<int>(<Object?>[r'PERSIST', key]);
+  Future<int> persist(K key) => runNonNull<int>(<Object>[r'PERSIST', key]);
 
   @override
   Future<int> pexpire(K key, int milliseconds) =>
-      run<int>(<Object?>[r'PEXPIRE', key, milliseconds]);
+      runNonNull<int>(<Object>[r'PEXPIRE', key, milliseconds]);
 
   @override
   Future<int> pexpireat(K key, int millisecondsTimestamp) =>
-      run<int>(<Object?>[r'PEXPIREAT', key, millisecondsTimestamp]);
+      runNonNull<int>(<Object>[r'PEXPIREAT', key, millisecondsTimestamp]);
 
   @override
-  Future<int> pttl(K key) => run<int>(<Object?>[r'PTTL', key]);
+  Future<int> pttl(K key) => runNonNull<int>(<Object>[r'PTTL', key]);
 
   @override
-  Future<K> randomkey() => run<K>(<Object>[r'RANDOMKEY']);
+  Future<K> randomkey() => runNonNull<K>(<Object>[r'RANDOMKEY']);
 
   @override
   Future<void> rename(K key, K newkey) =>
-      run<void>(<Object?>[r'RENAME', key, newkey]);
+      run<Never>(<Object?>[r'RENAME', key, newkey]);
 
   @override
   Future<int> renamenx(K key, K newkey) =>
-      run<int>(<Object?>[r'RENAMENX', key, newkey]);
+      runNonNull<int>(<Object>[r'RENAMENX', key, newkey]);
 
   @override
   Future<void> restore(K key, int ttl, List<int> serializedValue,
           {bool replace = false}) =>
-      run<void>(<Object?>[
+      run<Never>(<Object?>[
         r'RESTORE',
         key,
         ttl,
@@ -433,7 +434,7 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<KeyScanResult<K>> scan(int cursor, {K? pattern, int? count}) =>
-      run<KeyScanResult<K>>(<void>[
+      runNonNull<KeyScanResult<K>>(<Object?>[
         r'SCAN',
         cursor,
         pattern == null ? null : r'MATCH',
@@ -443,7 +444,7 @@ class Commands<K, V> extends ModuleBase
       ], mapper: KeyScanMapper<K>());
 
   @override
-  Future<List<V>> sort(K key,
+  Future<List<V?>> sort(K key,
       {K? by,
       int? offset,
       int? count,
@@ -453,7 +454,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<List<V>>(<Object?>[
+    return runNonNull<List<V?>>(<Object?>[
       r'SORT',
       key,
       by == null ? null : r'BY',
@@ -478,7 +479,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<int>(<Object?>[
+    return runNonNull<int>(<Object?>[
       r'SORT',
       key,
       by == null ? null : r'BY',
@@ -496,117 +497,118 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int> touch({K? key, Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'TOUCH', key, ...keys]);
+      runNonNull<int>(<Object?>[r'TOUCH', key, ...keys]);
 
   @override
-  Future<int> ttl(K key) => run<int>(<Object?>[r'TTL', key]);
+  Future<int> ttl(K key) => runNonNull<int>(<Object?>[r'TTL', key]);
 
   @override
-  Future<String> type(K key) => run<String>(<Object?>[r'TYPE', key]);
+  Future<String> type(K key) => runNonNull<String>(<Object?>[r'TYPE', key]);
 
   @override
   Future<int> unlink({K? key, Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'UNLINK', key, ...keys]);
+      runNonNull<int>(<Object?>[r'UNLINK', key, ...keys]);
 
   @override
   Future<int> wait(int numslaves, int timeout) =>
-      run<int>(<Object>[r'WAIT', numslaves, timeout]);
+      runNonNull<int>(<Object>[r'WAIT', numslaves, timeout]);
 
   // Lists.
 
   @override
   Future<ListPopResult<K?, V?>?> blpop(
           {K? key, Iterable<K?> keys = const [], int timeout = 0}) =>
-      run<ListPopResult<K?, V?>?>(<Object?>[r'BLPOP', key, ...keys, timeout],
+      run<ListPopResult<K?, V?>>(<Object?>[r'BLPOP', key, ...keys, timeout],
           mapper: ListPopResultMapper<K, V>());
 
   @override
   Future<ListPopResult<K?, V?>?> brpop(
           {K? key, Iterable<K?> keys = const [], int timeout = 0}) =>
-      run<ListPopResult<K?, V?>?>(<Object?>[r'BRPOP', key, ...keys, timeout],
+      run<ListPopResult<K?, V?>>(<Object?>[r'BRPOP', key, ...keys, timeout],
           mapper: ListPopResultMapper<K, V>());
 
   @override
   Future<V?> brpoplpush(K? source, K? destination, {int timeout = 0}) =>
-      run<V?>(<Object?>[r'BRPOPLPUSH', source, destination, timeout],
+      run<V>(<Object?>[r'BRPOPLPUSH', source, destination, timeout],
           mapper: BrpoplpushMapper<V>());
 
   @override
-  Future<V> lindex(K? key, int index) =>
+  Future<V?> lindex(K? key, int index) =>
       run<V>(<Object?>[r'LINDEX', key, index]);
 
   @override
   Future<int> linsert(K? key, InsertPosition position, V? pivot, V? value) =>
-      run<int>(<Object?>[r'LINSERT', key, position.name, pivot, value]);
+      runNonNull<int>(<Object?>[r'LINSERT', key, position.name, pivot, value]);
 
   @override
-  Future<int> llen(K? key) => run<int>(<Object?>[r'LLEN', key]);
+  Future<int> llen(K? key) => runNonNull<int>(<Object?>[r'LLEN', key]);
 
   @override
-  Future<V> lpop(K? key) => run<V>(<Object?>[r'LPOP', key]);
+  Future<V?> lpop(K? key) => run<V>(<Object?>[r'LPOP', key]);
 
   @override
   Future<int> lpush(K? key, {V? value, Iterable<V?> values = const []}) =>
-      run<int>(<Object?>[r'LPUSH', key, value, ...values]);
+      runNonNull<int>(<Object?>[r'LPUSH', key, value, ...values]);
 
   @override
   Future<int> lpushx(K? key, V? value) =>
-      run<int>(<Object?>[r'LPUSHX', key, value]);
+      runNonNull<int>(<Object?>[r'LPUSHX', key, value]);
 
   @override
   Future<List<V>> lrange(K? key, int start, int stop) =>
-      run<List<V>>(<Object?>[r'LRANGE', key, start, stop]);
+      runNonNull<List<V>>(<Object?>[r'LRANGE', key, start, stop]);
 
   @override
   Future<int> lrem(K? key, int count, V? value) =>
-      run<int>(<Object?>[r'LREM', key, count, value]);
+      runNonNull<int>(<Object?>[r'LREM', key, count, value]);
 
   @override
   Future<void> lset(K? key, int index, V? value) =>
-      run<void>(<Object?>[r'LSET', key, index, value]);
+      run<Never>(<Object?>[r'LSET', key, index, value]);
 
   @override
   Future<void> ltrim(K? key, int start, int stop) =>
-      run<void>(<Object?>[r'LTRIM', key, start, stop]);
+      run<Never>(<Object?>[r'LTRIM', key, start, stop]);
 
   @override
-  Future<V> rpop(K? key) => run<V>(<Object?>[r'RPOP', key]);
+  Future<V?> rpop(K? key) => run<V>(<Object?>[r'RPOP', key]);
 
   @override
-  Future<V> rpoplpush(K? source, K? destination) =>
+  Future<V?> rpoplpush(K? source, K? destination) =>
       run<V>(<Object?>[r'RPOPLPUSH', source, destination]);
 
   @override
   Future<int> rpush(K? key, {V? value, Iterable<V?> values = const []}) =>
-      run<int>(<Object?>[r'RPUSH', key, value, ...values]);
+      runNonNull<int>(<Object?>[r'RPUSH', key, value, ...values]);
 
   @override
   Future<int> rpushx(K? key, V? value) =>
-      run<int>(<Object?>[r'RPUSHX', key, value]);
+      runNonNull<int>(<Object?>[r'RPUSHX', key, value]);
 
   // Pub/Sub.
 
   @override
   Future<int> publish(K? channel, V message) =>
-      run<int>(<Object?>[r'PUBLISH', channel, message]);
+      runNonNull<int>(<Object?>[r'PUBLISH', channel, message]);
 
   @override
   Future<List<K>> pubsubChannels({K? pattern}) =>
-      run<List<K>>(<Object?>[r'PUBSUB', r'CHANNELS', pattern]);
+      runNonNull<List<K>>(<Object?>[r'PUBSUB', r'CHANNELS', pattern]);
 
   @override
   Future<List<PubsubResult<K?>>> pubsubNumsub(
           {Iterable<K?> channels = const []}) =>
-      run<List<PubsubResult<K?>>>(<Object?>[r'PUBSUB', r'NUMSUB', ...channels],
+      runNonNull<List<PubsubResult<K?>>>(
+          <Object?>[r'PUBSUB', r'NUMSUB', ...channels],
           mapper: PubsubResultMapper<K>());
 
   @override
-  Future<int> pubsubNumpat() => run<int>(<Object>[r'PUBSUB', r'NUMPAT']);
+  Future<int> pubsubNumpat() => runNonNull<int>(<Object>[r'PUBSUB', r'NUMPAT']);
 
   // Scripting.
 
   @override
-  Future<T> eval<T>(String script,
+  Future<T?> eval<T extends Object>(String script,
           {Iterable<K> keys = const [],
           Iterable<Object> args = const [],
           Mapper<T>? mapper}) =>
@@ -614,7 +616,7 @@ class Commands<K, V> extends ModuleBase
           mapper: mapper);
 
   @override
-  Future<T> evalsha<T>(String sha1,
+  Future<T?> evalsha<T extends Object>(String sha1,
           {Iterable<K> keys = const [],
           Iterable<Object> args = const [],
           Mapper<T>? mapper}) =>
@@ -623,39 +625,40 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<void> scriptDebug(ScriptDebugMode mode) =>
-      run<void>(<Object>[r'SCRIPT', r'DEBUG', mode.name]);
+      run<Never>(<Object>[r'SCRIPT', r'DEBUG', mode.name]);
 
   @override
   Future<List<int>> scriptExists(
           {String? sha1, Iterable<String> sha1s = const []}) =>
-      run<List<int>>(<Object?>[r'SCRIPT', r'EXISTS', sha1, ...sha1s]);
+      runNonNull<List<int>>(<Object?>[r'SCRIPT', r'EXISTS', sha1, ...sha1s]);
 
   @override
-  Future<void> scriptFlush() => run<void>(<Object>[r'SCRIPT', r'FLUSH']);
+  Future<void> scriptFlush() => run<Never>(<Object>[r'SCRIPT', r'FLUSH']);
 
   @override
-  Future<void> scriptKill() => run<void>(<Object>[r'SCRIPT', r'KILL']);
+  Future<void> scriptKill() => run<Never>(<Object>[r'SCRIPT', r'KILL']);
 
   @override
   Future<String> scriptLoad(String script) =>
-      run<String>(<Object>[r'SCRIPT', r'LOAD', script]);
+      runNonNull<String>(<Object>[r'SCRIPT', r'LOAD', script]);
 
   // Server.
 
   @override
-  Future<String> bgrewriteaof() => run<String>(<Object>[r'BGREWRITEAOF']);
+  Future<String> bgrewriteaof() =>
+      runNonNull<String>(<Object>[r'BGREWRITEAOF']);
 
   @override
-  Future<String> bgsave() => run<String>(<Object>[r'BGSAVE']);
+  Future<String> bgsave() => runNonNull<String>(<Object>[r'BGSAVE']);
 
   @override
-  Future<String> clientGetname() =>
+  Future<String?> clientGetname() =>
       run<String>(<Object>[r'CLIENT', r'GETNAME']);
 
   @override
   Future<int?> clientKill(
           {String? ipPort, Iterable<ClientFilter> filters = const []}) =>
-      run<int?>(<Object?>[
+      run<int>(<Object?>[
         r'CLIENT',
         r'KILL',
         ipPort,
@@ -675,105 +678,106 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<List<Map<String, String>>> clientList() =>
-      run<List<Map<String, String>>>(<Object>[r'CLIENT', r'LIST'],
+      runNonNull<List<Map<String, String>>>(<Object>[r'CLIENT', r'LIST'],
           mapper: clientListMapper);
 
   @override
   Future<void> clientPause(int timeout) =>
-      run<void>(<Object>[r'CLIENT', r'PAUSE', timeout]);
+      run<Never>(<Object>[r'CLIENT', r'PAUSE', timeout]);
 
   @override
   Future<void> clientReply(ReplyMode mode) {
-    final command = ClientReplyCommand<void>(
-        <Object>[r'CLIENT', r'REPLY', mode.name], mode);
+    final command =
+        ClientReplyCommand<Never>(<Object>[r'CLIENT', r'REPLY', mode.name], mode);
 
-    return execute(command);
+    return execute<Never>(command);
   }
 
   @override
   Future<void> clientSetname(String connectionName) =>
-      run<void>(<Object>[r'CLIENT', r'SETNAME', connectionName]);
+      run<Never>(<Object>[r'CLIENT', r'SETNAME', connectionName]);
 
   @override
   Future<List<ClientCommand?>> command() =>
-      run<List<ClientCommand?>>(<Object>[r'COMMAND'], mapper: commandMapper);
+      runNonNull<List<ClientCommand?>>(<Object>[r'COMMAND'],
+          mapper: commandMapper);
 
   @override
-  Future<int> commandCount() => run<int>(<Object>[r'COMMAND', r'COUNT']);
+  Future<int> commandCount() => runNonNull<int>(<Object>[r'COMMAND', r'COUNT']);
 
   @override
   Future<List<K>> commandGetkeys(List<Object> commandLine) =>
-      run<List<K>>(<Object>[r'COMMAND', r'GETKEYS', ...commandLine]);
+      runNonNull<List<K>>(<Object>[r'COMMAND', r'GETKEYS', ...commandLine]);
 
   @override
   Future<List<ClientCommand?>> commandInfo(
           {String? commandName, Iterable<String> commandNames = const []}) =>
-      run<List<ClientCommand?>>(
+      runNonNull<List<ClientCommand?>>(
           <Object?>[r'COMMAND', r'INFO', commandName, ...commandNames],
           mapper: commandMapper);
 
   @override
   Future<Map<String?, String?>> configGet(String parameter) =>
-      run<Map<String?, String?>>(<Object>[r'CONFIG', r'GET', parameter],
+      runNonNull<Map<String?, String?>>(<Object>[r'CONFIG', r'GET', parameter],
           mapper: configMapper);
 
   @override
-  Future<void> configResetstat() =>
-      run<void>(<Object>[r'CONFIG', r'RESETSTAT']);
+  Future<void> configResetstat() => run<Never>(<Object>[r'CONFIG', r'RESETSTAT']);
 
   @override
-  Future<void> configRewrite() => run<void>(<Object>[r'CONFIG', r'REWRITE']);
+  Future<void> configRewrite() => run<Never>(<Object>[r'CONFIG', r'REWRITE']);
 
   @override
   Future<void> configSet(String parameter, String value) =>
-      run<void>(<Object>[r'CONFIG', r'SET', parameter, value]);
+      run<Never>(<Object>[r'CONFIG', r'SET', parameter, value]);
 
   @override
-  Future<int> dbsize() => run<int>(<Object>[r'DBSIZE']);
+  Future<int> dbsize() => runNonNull<int>(<Object>[r'DBSIZE']);
 
   @override
   Future<String> debugObject(K key) =>
-      run<String>(<Object?>[r'DEBUG', r'OBJECT', key]);
+      runNonNull<String>(<Object?>[r'DEBUG', r'OBJECT', key]);
 
   @override
-  Future<void> debugSegfault() => run<void>(<Object>[r'DEBUG', r'SEGFAULT']);
+  Future<void> debugSegfault() => run<Never>(<Object>[r'DEBUG', r'SEGFAULT']);
 
   @override
   Future<void> flushall({bool asynchronously = false}) =>
-      run<void>(<Object?>[r'FLUSHALL', asynchronously ? r'ASYNC' : null]);
+      run<Never>(<Object?>[r'FLUSHALL', asynchronously ? r'ASYNC' : null]);
 
   @override
   Future<void> flushdb({bool asynchronously = false}) =>
-      run<void>(<Object?>[r'FLUSHDB', asynchronously ? r'ASYNC' : null]);
+      run<Never>(<Object?>[r'FLUSHDB', asynchronously ? r'ASYNC' : null]);
 
   @override
   Future<String> info([InfoSection? section]) =>
-      run<String>(<Object?>[r'INFO', section?.name]);
+      runNonNull<String>(<Object?>[r'INFO', section?.name]);
 
   @override
-  Future<int> lastsave() => run<int>(<Object>[r'LASTSAVE']);
+  Future<int> lastsave() => runNonNull<int>(<Object>[r'LASTSAVE']);
 
   @override
-  Future<String> memoryDoctor() => run<String>(<Object>[r'MEMORY', r'DOCTOR']);
+  Future<String> memoryDoctor() =>
+      runNonNull<String>(<Object>[r'MEMORY', r'DOCTOR']);
 
   @override
   Future<List<String>> memoryHelp() =>
-      run<List<String>>(<Object>[r'MEMORY', r'HELP']);
+      runNonNull<List<String>>(<Object>[r'MEMORY', r'HELP']);
 
   @override
   Future<String> memoryMallocStats() =>
-      run<String>(<Object>[r'MEMORY', r'MALLOC-STATS']);
+      runNonNull<String>(<Object>[r'MEMORY', r'MALLOC-STATS']);
 
   @override
-  Future<void> memoryPurge() => run<void>(<Object>[r'MEMORY', r'PURGE']);
+  Future<void> memoryPurge() => run<Never>(<Object>[r'MEMORY', r'PURGE']);
 
   @override
   Future<Map<String?, Object?>> memoryStats() =>
-      run<Map<String?, Object?>>(<Object>[r'MEMORY', r'STATS'],
+      runNonNull<Map<String?, Object?>>(<Object>[r'MEMORY', r'STATS'],
           mapper: memoryStatsMapper);
 
   @override
-  Future<int> memoryUsage(K key, {int? count}) => run<int>(<Object?>[
+  Future<int?> memoryUsage(K key, {int? count}) => run<int>(<Object?>[
         r'MEMORY',
         r'USAGE',
         key,
@@ -782,92 +786,94 @@ class Commands<K, V> extends ModuleBase
       ]);
 
   @override
-  Future<Role> role() => run<Role>(<Object>[r'ROLE'], mapper: roleMapper);
+  Future<Role> role() =>
+      runNonNull<Role>(<Object>[r'ROLE'], mapper: roleMapper);
 
   @override
-  Future<void> save() => run<void>(<Object>[r'SAVE']);
+  Future<void> save() => run<Never>(<Object>[r'SAVE']);
 
   @override
   Future<void> shutdown([ShutdownMode? mode]) =>
-      run<void>(<Object?>[r'SHUTDOWN', mode?.name]);
+      run<Never>(<Object?>[r'SHUTDOWN', mode?.name]);
 
   @override
   Future<void> slaveof(String host, String port) =>
-      run<void>(<Object>[r'SLAVEOF', host, port]);
+      run<Never>(<Object>[r'SLAVEOF', host, port]);
 
   @override
   Future<List<SlowLogEntry>> slowlogGet({int? count}) =>
-      run<List<SlowLogEntry>>(<Object?>[r'SLOWLOG', r'GET', count],
+      runNonNull<List<SlowLogEntry>>(<Object?>[r'SLOWLOG', r'GET', count],
           mapper: slowLogMapper);
 
   @override
-  Future<int> slowlogLen() => run<int>(<Object>[r'SLOWLOG', r'LEN']);
+  Future<int> slowlogLen() => runNonNull<int>(<Object>[r'SLOWLOG', r'LEN']);
 
   @override
-  Future<void> slowlogReset() => run<void>(<Object>[r'SLOWLOG', r'RESET']);
+  Future<void> slowlogReset() => run<Never>(<Object>[r'SLOWLOG', r'RESET']);
 
   @override
   Future<ServerTime> time() =>
-      run<ServerTime>(<Object>[r'TIME'], mapper: timeMapper);
+      runNonNull<ServerTime>(<Object>[r'TIME'], mapper: timeMapper);
 
   // Sets.
 
   @override
   Future<int> sadd(K key, {V? member, Iterable<V> members = const []}) =>
-      run<int>(<Object?>[r'SADD', key, member, ...members]);
+      runNonNull<int>(<Object?>[r'SADD', key, member, ...members]);
 
   @override
-  Future<int> scard(K key) => run<int>([r'SCARD', key]);
+  Future<int> scard(K key) => runNonNull<int>([r'SCARD', key]);
 
   @override
   Future<List<V>> sdiff(K key, {Iterable<K> keys = const []}) =>
-      run<List<V>>(<Object?>[r'SDIFF', key, ...keys]);
+      runNonNull<List<V>>(<Object?>[r'SDIFF', key, ...keys]);
 
   @override
   Future<int> sdiffstore(K destination, K key, {Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'SDIFFSTORE', destination, key, ...keys]);
+      runNonNull<int>(<Object?>[r'SDIFFSTORE', destination, key, ...keys]);
 
   @override
   Future<List<V>> sinter(K key, {Iterable<K> keys = const []}) =>
-      run<List<V>>(<Object?>[r'SINTER', key, ...keys]);
+      runNonNull<List<V>>(<Object?>[r'SINTER', key, ...keys]);
 
   @override
   Future<int> sinterstore(K destination, K key,
           {Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'SINTERSTORE', destination, key, ...keys]);
+      runNonNull<int>(<Object?>[r'SINTERSTORE', destination, key, ...keys]);
 
   @override
   Future<int> sismember(K key, V member) =>
-      run<int>(<Object?>[r'SISMEMBER', key, member]);
+      runNonNull<int>(<Object?>[r'SISMEMBER', key, member]);
 
   @override
-  Future<List<V>> smembers(K key) => run<List<V>>(<Object?>[r'SMEMBERS', key]);
+  Future<List<V>> smembers(K key) =>
+      runNonNull<List<V>>(<Object?>[r'SMEMBERS', key]);
 
   @override
   Future<int> smove(K source, K destination, V member) =>
-      run<int>(<Object?>[r'SMOVE', source, destination, member]);
+      runNonNull<int>(<Object?>[r'SMOVE', source, destination, member]);
 
   @override
-  Future<V> spop(K key) => run<V>(<Object?>[r'SPOP', key]);
+  Future<V?> spop(K key) => run<V>(<Object?>[r'SPOP', key]);
 
   @override
   Future<List<V>> spopCount(K key, int count) =>
-      run<List<V>>(<Object?>[r'SPOP', key, count]);
+      runNonNull<List<V>>(<Object?>[r'SPOP', key, count]);
 
   @override
-  Future<V> srandmember(K key) => run<V>(<Object?>[r'SRANDMEMBER', key]);
+  Future<V?> srandmember(K key) => run<V>(<Object?>[r'SRANDMEMBER', key]);
 
   @override
   Future<List<V>> srandmemberCount(K key, int count) =>
-      run<List<V>>(<Object?>[r'SRANDMEMBER', key, count]);
+      runNonNull<List<V>>(<Object?>[r'SRANDMEMBER', key, count]);
 
   @override
   Future<int> srem(K key, {V? member, Iterable<V> members = const []}) =>
-      run<int>(<Object?>[r'SREM', key, member, ...members]);
+      runNonNull<int>(<Object?>[r'SREM', key, member, ...members]);
 
   @override
   Future<SetScanResult<V>> sscan(K key, int cursor, {K? pattern, int? count}) =>
-      run<SetScanResult<V>>(<Object?>[
+      runNonNull<SetScanResult<V>>(<Object?>[
         r'SSCAN',
         key,
         cursor,
@@ -879,26 +885,26 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<List<V>> sunion(K key, {Iterable<K> keys = const []}) =>
-      run<List<V>>(<Object?>[r'SUNION', key, ...keys]);
+      runNonNull<List<V>>(<Object?>[r'SUNION', key, ...keys]);
 
   @override
   Future<int> sunionstore(K destination, K key,
           {Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'SUNIONSTORE', destination, key, ...keys]);
+      runNonNull<int>(<Object?>[r'SUNIONSTORE', destination, key, ...keys]);
 
   // Sorted sets.
 
   @override
   Future<SortedSetPopResult<K?, V?>?> bzpopmax(
           {K? key, Iterable<K?> keys = const [], int timeout = 0}) =>
-      run<SortedSetPopResult<K?, V?>?>(
+      run<SortedSetPopResult<K?, V?>>(
           <Object?>[r'BZPOPMAX', key, ...keys, timeout],
           mapper: SortedSetPopResultMapper<K, V>());
 
   @override
   Future<SortedSetPopResult<K?, V?>?> bzpopmin(
           {K? key, Iterable<K?> keys = const [], int timeout = 0}) =>
-      run<SortedSetPopResult<K?, V?>?>(
+      run<SortedSetPopResult<K?, V?>>(
           <Object?>[r'BZPOPMIN', key, ...keys, timeout],
           mapper: SortedSetPopResultMapper<K, V>());
 
@@ -909,7 +915,7 @@ class Commands<K, V> extends ModuleBase
           double? score,
           V? member,
           Map<V?, double> set = const {}}) =>
-      run<int>(<Object?>[
+      runNonNull<int>(<Object?>[
         r'ZADD',
         key,
         mode?.name,
@@ -920,25 +926,25 @@ class Commands<K, V> extends ModuleBase
       ]);
 
   @override
-  Future<double> zaddIncr(K? key, double score, V? member,
+  Future<double?> zaddIncr(K? key, double score, V? member,
           {SortedSetExistMode? mode}) =>
       run<double>(<Object?>[r'ZADD', key, mode?.name, r'INCR', score, member]);
 
   @override
-  Future<int> zcard(K? key) => run<int>(<Object?>[r'ZCARD', key]);
+  Future<int> zcard(K? key) => runNonNull<int>(<Object?>[r'ZCARD', key]);
 
   @override
   Future<int> zcount(K? key, String min, String max) =>
-      run<int>(<Object?>[r'ZCOUNT', key, min, max]);
+      runNonNull<int>(<Object?>[r'ZCOUNT', key, min, max]);
 
   @override
   Future<double> zincrby(K? key, double increment, V? member) =>
-      run<double>(<Object?>[r'ZINCRBY', key, increment, member]);
+      runNonNull<double>(<Object?>[r'ZINCRBY', key, increment, member]);
 
   @override
   Future<int> zinterstore(K? destination, List<K?> keys,
           {Iterable<double> weights = const [], AggregateMode? mode}) =>
-      run<int>(<Object?>[
+      runNonNull<int>(<Object?>[
         r'ZINTERSTORE',
         destination,
         keys.length,
@@ -951,22 +957,22 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int> zlexcount(K? key, V? min, V? max) =>
-      run<int>(<Object?>[r'ZLEXCOUNT', key, min, max]);
+      runNonNull<int>(<Object?>[r'ZLEXCOUNT', key, min, max]);
 
   @override
   Future<Map<V?, double?>?> zpopmax(K? key, {int? count}) =>
-      run<Map<V?, double?>?>(<Object?>[r'ZPOPMAX', key, count],
+      run<Map<V?, double?>>(<Object?>[r'ZPOPMAX', key, count],
           mapper: SortedSetMapper<V>(withScores: true));
 
   @override
   Future<Map<V?, double?>?> zpopmin(K? key, {int? count}) =>
-      run<Map<V?, double?>?>(<Object?>[r'ZPOPMIN', key, count],
+      run<Map<V?, double?>>(<Object?>[r'ZPOPMIN', key, count],
           mapper: SortedSetMapper<V>(withScores: true));
 
   @override
   Future<Map<V?, double?>?> zrange(K? key, int start, int stop,
           {bool withScores = false}) =>
-      run<Map<V?, double?>?>(<Object?>[
+      run<Map<V?, double?>>(<Object?>[
         r'ZRANGE',
         key,
         start,
@@ -980,7 +986,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<List<V>>(<Object?>[
+    return runNonNull<List<V>>(<Object?>[
       r'ZRANGEBYLEX',
       key,
       min,
@@ -997,7 +1003,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<Map<V?, double?>?>(<Object?>[
+    return run<Map<V?, double?>>(<Object?>[
       r'ZRANGEBYSCORE',
       key,
       min,
@@ -1010,29 +1016,29 @@ class Commands<K, V> extends ModuleBase
   }
 
   @override
-  Future<int> zrank(K? key, V? member) =>
+  Future<int?> zrank(K? key, V? member) =>
       run<int>(<Object?>[r'ZRANK', key, member]);
 
   @override
   Future<int> zrem(K? key, {V? member, Iterable<V?> members = const []}) =>
-      run<int>(<Object?>[r'ZREM', key, member, ...members]);
+      runNonNull<int>(<Object?>[r'ZREM', key, member, ...members]);
 
   @override
   Future<int> zremrangebylex(K? key, V? min, V? max) =>
-      run<int>(<Object?>[r'ZREMRANGEBYLEX', key, min, max]);
+      runNonNull<int>(<Object?>[r'ZREMRANGEBYLEX', key, min, max]);
 
   @override
   Future<int> zremrangebyrank(K? key, int start, int stop) =>
-      run<int>(<Object?>[r'ZREMRANGEBYRANK', key, start, stop]);
+      runNonNull<int>(<Object?>[r'ZREMRANGEBYRANK', key, start, stop]);
 
   @override
   Future<int> zremrangebyscore(K? key, String min, String max) =>
-      run<int>(<Object?>[r'ZREMRANGEBYSCORE', key, min, max]);
+      runNonNull<int>(<Object?>[r'ZREMRANGEBYSCORE', key, min, max]);
 
   @override
   Future<Map<V?, double?>?> zrevrange(K? key, int start, int stop,
           {bool withScores = false}) =>
-      run<Map<V?, double?>?>(<Object?>[
+      run<Map<V?, double?>>(<Object?>[
         r'ZREVRANGE',
         key,
         start,
@@ -1046,7 +1052,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<List<V>>(<Object?>[
+    return runNonNull<List<V>>(<Object?>[
       r'ZREVRANGEBYLEX',
       key,
       max,
@@ -1063,7 +1069,7 @@ class Commands<K, V> extends ModuleBase
     assert(
         (offset == null && count == null) || (offset != null && count != null));
 
-    return run<Map<V?, double?>?>(<Object?>[
+    return run<Map<V?, double?>>(<Object?>[
       r'ZREVRANGEBYSCORE',
       key,
       max,
@@ -1076,13 +1082,13 @@ class Commands<K, V> extends ModuleBase
   }
 
   @override
-  Future<int> zrevrank(K? key, V? member) =>
+  Future<int?> zrevrank(K? key, V? member) =>
       run<int>(<Object?>[r'ZREVRANK', key, member]);
 
   @override
-  Future<SortedSetScanResult<K?>> zscan(K? key, int cursor,
+  Future<SortedSetScanResult<K>> zscan(K? key, int cursor,
           {K? pattern, int? count}) =>
-      run<SortedSetScanResult<K?>>(<Object?>[
+      runNonNull<SortedSetScanResult<K>>(<Object?>[
         r'ZSCAN',
         key,
         cursor,
@@ -1093,13 +1099,13 @@ class Commands<K, V> extends ModuleBase
       ], mapper: SortedSetScanMapper<K>());
 
   @override
-  Future<double> zscore(K? key, V? member) =>
+  Future<double?> zscore(K? key, V? member) =>
       run<double>(<Object?>[r'ZSCORE', key, member]);
 
   @override
   Future<int> zunionstore(K? destination, List<K?> keys,
           {Iterable<double> weights = const [], AggregateMode? mode}) =>
-      run<int>(<Object?>[
+      runNonNull<int>(<Object?>[
         r'ZUNIONSTORE',
         destination,
         keys.length,
@@ -1114,7 +1120,7 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int> xack(K? key, K? group, {K? id, Iterable<K?> ids = const []}) =>
-      run<int>(<Object?>[r'XACK', key, group, id, ...ids]);
+      runNonNull<int>(<Object?>[r'XACK', key, group, id, ...ids]);
 
   @override
   Future<K> xadd(K? key,
@@ -1124,7 +1130,7 @@ class Commands<K, V> extends ModuleBase
           Map<K?, V?> fields = const {},
           int? maxlen,
           bool roughly = false}) =>
-      run<K>(<Object?>[
+      runNonNull<K>(<Object?>[
         r'XADD',
         key,
         maxlen == null ? null : r'MAXLEN',
@@ -1145,7 +1151,7 @@ class Commands<K, V> extends ModuleBase
           int? retryCount,
           bool force = false,
           bool justId = false}) =>
-      run<Object>(<Object?>[
+      runNonNull<Object>(<Object?>[
         r'XCLAIM',
         key,
         group,
@@ -1165,12 +1171,12 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int> xdel(K? key, {K? id, Iterable<K?> ids = const []}) =>
-      run<int>(<Object?>[r'XDEL', key, id, ...ids]);
+      runNonNull<int>(<Object?>[r'XDEL', key, id, ...ids]);
 
   @override
   Future<Object?> xgroup(StreamGroupSubcommand subcommand,
           {K? key, K? group, K? id, K? consumer, bool mkstream = false}) =>
-      run<Object?>(<Object?>[
+      run<Object>(<Object?>[
         r'XGROUP',
         subcommand.name,
         key,
@@ -1182,11 +1188,11 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<Object> xinfo(StreamInfoSubcommand subcommand, {K? key, K? group}) =>
-      run<Object>(<Object?>[r'XINFO', subcommand.name, key, group],
+      runNonNull<Object>(<Object?>[r'XINFO', subcommand.name, key, group],
           mapper: StreamInfoMapper<K, V>(subcommand));
 
   @override
-  Future<int> xlen(K? key) => run<int>(<Object?>[r'XLEN', key]);
+  Future<int> xlen(K? key) => runNonNull<int>(<Object?>[r'XLEN', key]);
 
   @override
   Future<Object> xpending(K? key, K? group,
@@ -1195,7 +1201,7 @@ class Commands<K, V> extends ModuleBase
         (start != null && end != null && count != null) ||
         (start != null && end != null && count != null && consumer != null));
 
-    return run<Object>(
+    return runNonNull<Object>(
         <Object?>[r'XPENDING', key, group, start, end, count, consumer],
         mapper: StreamPendingMapper<K, V>(justSummary: start == null));
   }
@@ -1203,7 +1209,7 @@ class Commands<K, V> extends ModuleBase
   @override
   Future<List<StreamEntry<K?, V?>?>> xrange(K? key, K? start, K? end,
           {int? count}) =>
-      run<List<StreamEntry<K?, V?>?>>(<Object?>[
+      runNonNull<List<StreamEntry<K?, V?>?>>(<Object?>[
         r'XRANGE',
         key,
         start,
@@ -1220,7 +1226,7 @@ class Commands<K, V> extends ModuleBase
           Iterable<K?> ids = const [],
           int? count,
           int? timeout}) =>
-      run<Map<K?, List<StreamEntry<K?, V?>?>>?>(<Object?>[
+      run<Map<K?, List<StreamEntry<K?, V?>?>>>(<Object?>[
         r'XREAD',
         count == null ? null : r'COUNT',
         count,
@@ -1242,7 +1248,7 @@ class Commands<K, V> extends ModuleBase
           int? count,
           int? timeout,
           bool noack = false}) =>
-      run<Map<K?, List<StreamEntry<K?, V?>?>>?>(<Object?>[
+      run<Map<K?, List<StreamEntry<K?, V?>?>>>(<Object?>[
         r'XREADGROUP',
         r'GROUP',
         group,
@@ -1262,7 +1268,7 @@ class Commands<K, V> extends ModuleBase
   @override
   Future<List<StreamEntry<K?, V?>?>> xrevrange(K? key, K? end, K? start,
           {int? count}) =>
-      run<List<StreamEntry<K?, V?>?>>(<Object?>[
+      runNonNull<List<StreamEntry<K?, V?>?>>(<Object?>[
         r'XREVRANGE',
         key,
         end,
@@ -1272,24 +1278,25 @@ class Commands<K, V> extends ModuleBase
       ], mapper: StreamMapper<K, V>());
 
   @override
-  Future<int> xtrim(K? key, int maxlen, {bool roughly = false}) => run<int>(
-      <Object?>[r'XTRIM', key, r'MAXLEN', roughly ? r'~' : null, maxlen]);
+  Future<int> xtrim(K? key, int maxlen, {bool roughly = false}) =>
+      runNonNull<int>(
+          <Object?>[r'XTRIM', key, r'MAXLEN', roughly ? r'~' : null, maxlen]);
 
   // Strings.
 
   @override
   Future<int> append(K key, V value) =>
-      run<int>(<Object?>[r'APPEND', key, value]);
+      runNonNull<int>(<Object?>[r'APPEND', key, value]);
 
   @override
   Future<int> bitcount(K key, [int? start, int? end]) {
     assert((start == null && end == null) || (start != null && end != null));
-    return run<int>(<Object?>[r'BITCOUNT', key, start, end]);
+    return runNonNull<int>(<Object?>[r'BITCOUNT', key, start, end]);
   }
 
   @override
-  Future<List<int>> bitfield(K key, List<BitfieldOperation> operations) =>
-      run<List<int>>(
+  Future<List<int?>> bitfield(K key, List<BitfieldOperation> operations) =>
+      runNonNull<List<int?>>(
           <Object?>[r'BITFIELD', key, ...operations.expand(_expandBitfield)]);
 
   List<Object?> _expandBitfield(BitfieldOperation operation) => <Object?>[
@@ -1304,51 +1311,53 @@ class Commands<K, V> extends ModuleBase
   @override
   Future<int> bitop(BitopOperation operation, K destkey,
           {K? key, Iterable<K> keys = const []}) =>
-      run<int>(<Object?>[r'BITOP', operation.name, destkey, key, ...keys]);
+      runNonNull<int>(
+          <Object?>[r'BITOP', operation.name, destkey, key, ...keys]);
 
   @override
   Future<int> bitpos(K key, int bit, [int? start, int? end]) =>
-      run<int>(<Object?>[r'BITPOS', key, bit, start, end]);
+      runNonNull<int>(<Object?>[r'BITPOS', key, bit, start, end]);
 
   @override
-  Future<int> decr(K key) => run<int>(<Object?>[r'DECR', key]);
+  Future<int> decr(K key) => runNonNull<int>(<Object?>[r'DECR', key]);
 
   @override
   Future<int> decrby(K key, int decrement) =>
-      run<int>(<Object?>[r'DECRBY', key, decrement]);
+      runNonNull<int>(<Object?>[r'DECRBY', key, decrement]);
 
   @override
-  Future<V> get(K key) => run<V>(<Object?>[r'GET', key]);
+  Future<V?> get(K key) => run<V>(<Object?>[r'GET', key]);
 
   @override
   Future<int> getbit(K key, int offset) =>
-      run<int>(<Object?>[r'GETBIT', key, offset]);
+      runNonNull<int>(<Object?>[r'GETBIT', key, offset]);
 
   @override
   Future<V> getrange(K key, int start, int end) =>
-      run<V>(<Object?>[r'GETRANGE', key, start, end]);
+      runNonNull<V>(<Object?>[r'GETRANGE', key, start, end]);
 
   @override
-  Future<V> getset(K key, V value) => run<V>(<Object?>[r'GETSET', key, value]);
+  Future<V?> getset(K key, V value) =>
+      run<V>(<Object?>[r'GETSET', key, value]);
 
   @override
-  Future<int> incr(K key) => run<int>(<Object?>[r'INCR', key]);
+  Future<int> incr(K key) => runNonNull<int>(<Object?>[r'INCR', key]);
 
   @override
   Future<int> incrby(K key, int increment) =>
-      run<int>(<Object?>[r'INCRBY', key, increment]);
+      runNonNull<int>(<Object?>[r'INCRBY', key, increment]);
 
   @override
   Future<double> incrbyfloat(K key, double increment) =>
-      run<double>(<Object?>[r'INCRBYFLOAT', key, increment]);
+      runNonNull<double>(<Object?>[r'INCRBYFLOAT', key, increment]);
 
   @override
-  Future<List<V>> mget({K? key, Iterable<K> keys = const []}) =>
-      run<List<V>>(<Object?>[r'MGET', key, ...keys]);
+  Future<List<V?>> mget({K? key, Iterable<K> keys = const []}) =>
+      runNonNull<List<V?>>(<Object?>[r'MGET', key, ...keys]);
 
   @override
   Future<void> mset({K? key, V? value, Map<K, V> map = const {}}) =>
-      run<void>(<Object?>[
+      run<Never>(<Object?>[
         r'MSET',
         key,
         value,
@@ -1357,7 +1366,7 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<int> msetnx({K? key, V? value, Map<K, V> map = const {}}) =>
-      run<int>(<Object?>[
+      runNonNull<int>(<Object?>[
         r'MSETNX',
         key,
         value,
@@ -1366,16 +1375,16 @@ class Commands<K, V> extends ModuleBase
 
   @override
   Future<void> psetex(K key, int milliseconds, V value) =>
-      run<void>(<Object?>[r'PSETEX', key, milliseconds, value]);
+      run<Never>(<Object?>[r'PSETEX', key, milliseconds, value]);
 
   @override
   Future<bool> set(K key, V value,
-      {int? seconds, int? milliseconds, SetExistMode? mode}) {
+      {int? seconds, int? milliseconds, SetExistMode? mode}) async {
     assert((seconds == null && milliseconds == null) ||
         (seconds != null && milliseconds == null) ||
         (seconds == null && milliseconds != null));
 
-    return run<bool>(<Object?>[
+    return await run<bool>(<Object?>[
       r'SET',
       key,
       value,
@@ -1384,27 +1393,27 @@ class Commands<K, V> extends ModuleBase
       milliseconds == null ? null : r'PX',
       milliseconds,
       mode?.name
-    ], mapper: stringSetMapper);
+    ], mapper: stringSetMapper) ?? false;
   }
 
   @override
   Future<int> setbit(K key, int offset, int value) =>
-      run<int>(<Object?>[r'SETBIT', key, offset, value]);
+      runNonNull<int>(<Object?>[r'SETBIT', key, offset, value]);
 
   @override
   Future<void> setex(K key, int seconds, V value) =>
-      run<void>(<Object?>[r'SETEX', key, seconds, value]);
+      run<Never>(<Object?>[r'SETEX', key, seconds, value]);
 
   @override
   Future<int> setnx(K key, V value) =>
-      run<int>(<Object?>[r'SETNX', key, value]);
+      runNonNull<int>(<Object?>[r'SETNX', key, value]);
 
   @override
   Future<int> setrange(K key, int offset, V value) =>
-      run<int>(<Object?>[r'SETRANGE', key, offset, value]);
+      runNonNull<int>(<Object?>[r'SETRANGE', key, offset, value]);
 
   @override
-  Future<int> strlen(K key) => run<int>(<Object?>[r'STRLEN', key]);
+  Future<int> strlen(K key) => runNonNull<int>(<Object?>[r'STRLEN', key]);
 
   // Transactions.
 
@@ -1412,27 +1421,27 @@ class Commands<K, V> extends ModuleBase
   Future<void> discard() {
     final command = DiscardCommand(<Object>[r'DISCARD']);
 
-    return execute(command);
+    return execute<Never>(command);
   }
 
   @override
   Future<void> exec() {
     final command = ExecCommand(<Object>[r'EXEC']);
 
-    return execute(command);
+    return execute<Never>(command);
   }
 
   @override
   Future<void> multi() {
     final command = MultiCommand(<Object>[r'MULTI']);
 
-    return execute(command);
+    return execute<Never>(command);
   }
 
   @override
-  Future<void> unwatch() => run<void>(<Object>[r'UNWATCH']);
+  Future<void> unwatch() => run<Never>(<Object>[r'UNWATCH']);
 
   @override
   Future<void> watch({K? key, Iterable<K> keys = const []}) =>
-      run<void>(<Object?>[r'WATCH', key, ...keys]);
+      run<Never>(<Object?>[r'WATCH', key, ...keys]);
 }
