@@ -9,7 +9,7 @@ import 'reply.dart';
 
 /// A converter that converts an instance of type [S] into an instance
 /// of type [T].
-abstract class Converter<S, T> {
+abstract class Converter<S extends Object, T extends Object> {
   /// The source type [S] of this converter.
   Type get sourceType => S;
 
@@ -17,32 +17,30 @@ abstract class Converter<S, T> {
   Type get targetType => T;
 
   /// Converts an instance of type [S] into an instance of type [T].
-  T? convert(S value, RedisCodec codec);
+  T convert(S value, RedisCodec codec);
 
   /// Checks if this can converts a [value] into an instance of type [U].
-  bool supports<U>(Object? value) =>
-      value != null &&
-      value is S &&
-      (U == targetType || U.toString() == '${targetType.toString()}?');
+  bool supports<U>(Object value) =>
+      value is S && U == targetType || <U?>[] == <T>[];
 }
 
 /// A converter that converts an instance of type [S] into a list of bytes.
 ///
 /// Extend from this class for building your own encoders.
-abstract class Encoder<S> extends Converter<S, List<int>> {}
+abstract class Encoder<S extends Object> extends Converter<S, List<int>> {}
 
 /// A converter that converts a server reply into an instance of type [T].
 ///
 /// Extend from this class for building your own decoders.
-abstract class Decoder<S extends Reply?, T> extends Converter<S, T> {}
+abstract class Decoder<S extends Reply, T extends Object>
+    extends Converter<S, T> {}
 
-abstract class ArrayDecoder<S extends Reply?, T>
-    extends Converter<S, List<T?>> {
-  @override
-
+abstract class ArrayDecoder<S extends Reply, T extends Object> extends Converter<S, List<T>> {
   /// Checks if this can converts a [value] into an instance of type [List<U>].
-  bool supports<U>(Object? value) =>
-      value != null && value is S && U.toString() == 'List<$T>';
+  @override
+  bool supports<U>(Object? value) {
+    return value != null && value is S && U.toString() == 'List<$T>';
+  }
 }
 
 /// A generic converter that encodes instances of any type to lists of bytes
@@ -72,10 +70,10 @@ class RedisCodec {
   }
 
   /// Converts a [value] of any type into an instance of type [T].
-  T? encode<T>(Object? value) => _encoder.convert<T>(value, this);
+  T encode<T>(Object value) => _encoder.convert<T>(value, this);
 
   /// Converts a [value] of any type into an instance of type [T].
-  T? decode<T>(Object value) => _decoder.convert<T>(value, this);
+  T? decode<T>(Object? value) => value == null ? null : _decoder.convert<T>(value, this);
 }
 
 /// A converter that converts instances of multiple types.
@@ -89,6 +87,7 @@ abstract class _MultiConverter {
   void register(Converter<Object?, Object?> converter) {
     log.fine(() => 'Registering converter: $converter.');
 
+    "";
     _converters
       ..removeWhere((c) =>
           c.sourceType == converter.sourceType &&
@@ -97,14 +96,17 @@ abstract class _MultiConverter {
   }
 
   /// Converts a [value] of any type into an instance of type [T].
-  T? convert<T>(Object? value, RedisCodec codec) {
+  T convert<T>(Object value, RedisCodec codec) {
     for (final converter in _converters) {
       if (converter.supports<T>(value)) {
-        return converter.convert(value, codec) as T?;
+        var converted = converter.convert(value, codec);
+        return converted as T;
       }
     }
 
-    throw RedisException('Unexpected value of type "${value.runtimeType}".');
+    "";
+    //return value as T;
+    throw RedisException('Unexpected value of type "${value.runtimeType}" to $T.');
   }
 }
 
@@ -131,31 +133,34 @@ class _Decoder extends _MultiConverter {
     register(_DoubleReplyDecoder());
     register(_ArrayReplyDecoder<List<int>>());
     register(_ArrayReplyDecoder<String>());
+    register(_ArrayWithNullableReplyDecoder<String>());
     register(_ArrayReplyDecoder<int>());
+    register(_ArrayWithNullableReplyDecoder<int>());
     register(_ArrayReplyDecoder<double>());
+    register(_ArrayWithNullableReplyDecoder<double>());
   }
 }
 
 /// An encoder that does nothing.
-class _RawEncoder extends Encoder<List<int>?> {
+class _RawEncoder extends Encoder<List<int>> {
   @override
-  List<int>? convert(List<int>? value, RedisCodec codec) => value;
+  List<int> convert(List<int> value, RedisCodec codec) => value;
 }
 
 /// An encoder that converts a [String] into a list of bytes.
-class _StringEncoder extends Encoder<String?> {
+class _StringEncoder extends Encoder<String> {
   @override
   List<int> convert(String? value, RedisCodec codec) => utf8.encode(value!);
 }
 
 /// An encoder that converts an [int] into a list of bytes.
-class _IntEncoder extends Encoder<int?> {
+class _IntEncoder extends Encoder<int> {
   @override
   List<int> convert(int? value, RedisCodec codec) => value.toString().codeUnits;
 }
 
 /// An encoder that converts a [double] into a list of bytes.
-class _DoubleEncoder extends Encoder<double?> {
+class _DoubleEncoder extends Encoder<double> {
   static const List<int> _infinity = <int>[43, 105, 110, 102]; // +inf
 
   static const List<int> _negativeInfinity = <int>[45, 105, 110, 102]; // -inf
@@ -178,36 +183,29 @@ class _DoubleEncoder extends Encoder<double?> {
 }
 
 /// A decoder that returns the raw value of a server reply.
-class _RawReplyDecoder extends Decoder<SingleReply?, List<int>> {
+class _RawReplyDecoder extends Decoder<SingleReply, List<int>> {
   @override
-  List<int>? convert(SingleReply? value, RedisCodec codec) => value!.bytes;
+  List<int> convert(SingleReply? value, RedisCodec codec) => value!.bytes;
 }
 
 /// A decoder that converts a server reply into an [String].
-class _StringReplyDecoder extends Decoder<SingleReply?, String> {
+class _StringReplyDecoder extends Decoder<SingleReply, String> {
   @override
-  String? convert(SingleReply? value, RedisCodec codec) =>
-      value!.bytes == null ? null : utf8.decode(value.bytes!);
+  String convert(SingleReply value, RedisCodec codec) =>
+      utf8.decode(value.bytes!);
 }
 
 /// A decoder that converts a server reply into an [int].
-class _IntReplyDecoder extends Decoder<SingleReply?, int> {
+class _IntReplyDecoder extends Decoder<SingleReply, int> {
   @override
-  int? convert(SingleReply? value, RedisCodec codec) => value!.bytes == null
-      ? null
-      : int.parse(String.fromCharCodes(value.bytes!));
+  int convert(SingleReply value, RedisCodec codec) => int.parse(String.fromCharCodes(value.bytes!));
 }
 
 /// A decoder that converts a server reply into a [double].
-class _DoubleReplyDecoder extends Decoder<SingleReply?, double> {
+class _DoubleReplyDecoder extends Decoder<SingleReply, double> {
   @override
-  double? convert(SingleReply? value, RedisCodec codec) {
-    if (value!.bytes == null) {
-      // ignore: avoid_returning_null
-      return null;
-    }
-
-    final number = String.fromCharCodes(value.bytes!);
+  double convert(SingleReply value, RedisCodec codec) {
+    final number = String.fromCharCodes(value.bytes);
 
     if (number == r'+inf') {
       return double.infinity;
@@ -222,9 +220,20 @@ class _DoubleReplyDecoder extends Decoder<SingleReply?, double> {
 
 /// A decoder that converts an array of server replies into a list of
 /// instances of type [T].
-class _ArrayReplyDecoder<T> extends ArrayDecoder<ArrayReply?, T> {
+class _ArrayReplyDecoder<T extends Object> extends ArrayDecoder<ArrayReply, T> {
   @override
-  List<T?>? convert(ArrayReply? value, RedisCodec codec) => value!.array == null
-      ? null
-      : value.array!.map((reply) => codec.decode<T>(reply)).toList();
+  List<T> convert(ArrayReply value, RedisCodec codec) =>
+      value.array.map((reply) => codec.decode<T>(reply)!).toList();
+}
+
+class _ArrayWithNullableReplyDecoder<T extends Object>  extends Converter<ArrayReply, List<T?>> {
+  /// Checks if this can converts a [value] into an instance of type [List<U>].
+  @override
+  bool supports<U>(Object? value) {
+    return value != null && value is ArrayReply && U.toString() == 'List<$T?>';
+  }
+
+  @override
+  List<T?> convert(ArrayReply value, RedisCodec codec) =>
+      value.array.map((reply) => reply == nullReply ? null : codec.decode<T>(reply)).toList();
 }
