@@ -1,8 +1,6 @@
 // Copyright (c) 2018, Juan Mellado. All rights reserved. Use of this source
 // is governed by a MIT-style license that can be found in the LICENSE file.
 
-// @dart=2.9
-
 import 'package:test/test.dart';
 
 // ignore: directives_ordering
@@ -158,7 +156,7 @@ void main() {
 
       // EXEC
       final exec = ExecCommand(<Object>['EXEC']);
-      const reply = BulkReply(null);
+      const reply = BulkReply([]);
       transaction.onReply(exec, reply, codec);
 
       expect(set.future, throwRedisException);
